Lab3 Update: add DTO

+++ src/logic/roomController.cpp
@@ -9,7 +9,7 @@ RoomController::~RoomController(){}
 
 void RoomController::addRoom(std::string roomType, int number)
 {
-    this->Repository->addRoom(roomType, number);
+    this->Repository->addRoom(RoomDTO(roomType, number));
 }
 
 std::vector<Room> RoomController::getRooms()
@@ -17,14 +17,16 @@ std::vector<Room> RoomController::getRooms()
     return this->Repository->getRooms();
 }
 
-Room RoomController::getRoomInfo(int id)
+Room RoomController::getRoom(int id)
 {
-    return this->Repository->getRoomInfo(id);
+    return this->Repository->getRoom(id);
 }
 
 void RoomController::deleteRoom(int id)
 {
-    this->Repository->deleteRoom(id);
+    Room tmpRoom = this->Repository->getRoom(id);
+    if (tmpRoom.getID() >= 0)
+        this->Repository->deleteRoom(id);
 }
 
 std::vector<Thing> RoomController::getRoomThings(int id)
